CSuperSample Class

This class represents a collection of CSample objects (samples of connectomes), providing methods to aggregate, analyze, and compute statistics across multiple samples sharing the same Riemannian metric.

Active bindings

list_of_samples

The list of CSample objects aggregated in this super-sample.

sample_size

The total number of connectomes in all samples.

matrix_size

The size of the connectome matrices.

mfd_dim

The dimension of the manifold.

riem_metric

The Riemannian metric used by all samples.

variation

The total variation of the aggregated sample.

sample_cov

The sample covariance matrix of the aggregated sample.

full_sample

The aggregated CSample object containing all connectomes.

frechet_mean

The Frechet mean of the aggregated sample.

Within

The within-group covariance matrix (W).

Total

The total covariance matrix (T).

Methods

Public methods

CSuperSample$new()

CSuperSample$compute_variation()

CSuperSample$compute_sample_cov()

CSuperSample$gather()

CSuperSample$compute_fmean()

CSuperSample$compute_W()

CSuperSample$compute_T()

CSuperSample$clone()

Method new()

Initialize a CSuperSample object

Usage

CSuperSample$new(samples)

Arguments

samples

A list of CSample objects. All must use the same Riemannian metric.

Returns

A new CSuperSample object.

Method compute_variation()

Compute the total variation of the aggregated sample.

Usage

CSuperSample$compute_variation()

Returns

None. This function is called for its side effects. The result is stored in the variation active binding.

Method compute_sample_cov()

Compute the sample covariance matrix of the aggregated sample.

Usage

CSuperSample$compute_sample_cov()

Returns

None. This function is called for its side effects. The result is stored in the sample_cov active binding.

Method gather()

Gather all connectomes from the list of samples into a single CSample object.

Usage

CSuperSample$gather()

Returns

None. This function is called for its side effects. The result is stored in the full_sample active binding.

Method compute_fmean()

Compute the Frechet mean of the aggregated sample.

Usage

CSuperSample$compute_fmean()

Returns

None. This function is called for its side effects. The result is stored in the frechet_mean active binding.

Method compute_W()

Compute the within-group covariance matrix (W) for the samples.

Usage

CSuperSample$compute_W()

Returns

None. This function is called for its side effects. The result is stored in the Within active binding.

Method compute_T()

Compute the total covariance matrix (T) for the samples.

Usage

CSuperSample$compute_T()

Returns

None. This function is called for its side effects. The result is stored in the Total active binding.

Method clone()

The objects of this class are cloneable with this method.

Usage

CSuperSample$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.