riemtan: Riemannian Metrics for Symmetric Positive Definite Matrices

Implements various Riemannian metrics for symmetric positive definite matrices, including AIRM (Affine Invariant Riemannian Metric, doi:10.1007/s11263-005-3222-z ), Log-Euclidean (doi:10.1002/mrm.20965 ), Euclidean, Log-Cholesky (doi:10.1137/18M1221084 ), and Bures-Wasserstein metrics (doi:10.1016/j.exmath.2018.01.002 ). Provides functions for computing logarithmic and exponential maps, vectorization, and statistical operations on the manifold of positive definite matrices.

See also

Author

Maintainer: Nicolas Escobar nescoba@iu.edu (ORCID)

Other contributors:

Jaroslaw Harezlak [thesis advisor]