DataBackend Abstract Class

Abstract base class defining the interface for storage backends. All backend implementations must inherit from this class and implement the required methods.

Details

This class provides a common interface for different storage strategies:

ListBackend: In-memory list storage (current default)

ParquetBackend: Lazy-loaded Parquet files with LRU cache

Methods

Public methods

DataBackend$get_matrix()

DataBackend$get_all_matrices()

DataBackend$length()

DataBackend$get_dimensions()

DataBackend$clone()

Method get_matrix()

Get a specific matrix by index

Usage

DataBackend$get_matrix(i)

Arguments

i

Integer index of the matrix to retrieve

Returns

A dppMatrix object

Method get_all_matrices()

Get all matrices

Usage

DataBackend$get_all_matrices()

Returns

A list of dppMatrix objects

Method length()

Get the number of matrices

Usage

DataBackend$length()

Returns

Integer count of matrices

Method get_dimensions()

Get matrix dimensions

Usage

DataBackend$get_dimensions()

Returns

Integer dimension (p) where matrices are p x p

Method clone()

The objects of this class are cloneable with this method.

Usage

DataBackend$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.